feat: user profile page nearly done, add zustand for global auth store, misc. work
This commits adds:

A nearly complete user profile page. It currently pulls from dummy data,
since much of the user profile page is dependant on the level editor
being able to uplaod levels

The addition of Zustand for global stores. Storing auth info in a global
store enabled me to conditional rendering in components much more easily

Added custom error page route. Added delete level  button to LevelModal
that only renders if the logged in user is the same as the level author.