feat: level thumbnails get generated and stored
This commit adds the feature of level thumbnails automatically being
generated and uploaded to R2. The thumbnails are generated by a Python
script that uses the Pillow library. It takes in the level's peg layout
and pastes the pegs accordingly. The script is spawned as a child
process in TypeScript.